env/Cleanup-envrionment: Changed directory structure and code format
