+++ .clang-format
@@ -1,0 +1,3 @@
+---
+
+BasedOnStyle: Microsoft
