+++ .gitignore
@@ -52,3 +52,13 @@ modules.order
 Module.symvers
 Mkfile.old
 dkms.conf
+
+### Python
+.venv
+
+### Precommit
+.pre-commit-config.yaml
+
+### VS Code
+.vscode
+.devcontainer
